+++ inventory-management/src/pages/sales/SalesFilters.tsx
@@ -1,0 +1,241 @@
+import React from "react";
+import {
+  Filter,
+  FileText,
+  RefreshCw,
+  Download,
+  Plus,
+  ChevronDown,
+  ChevronUp
+} from "lucide-react";
+
+interface SalesFiltersProps {
+  showFilters: boolean;
+  setShowFilters: (show: boolean) => void;
+  filters: any;
+  handleFilterChange: (e: React.ChangeEvent<HTMLSelectElement | HTMLInputElement>) => void;
+  products: any[];
+  salers: any[];
+  clients: any[];
+  fetchSales: () => void;
+  viewType: "table" | "cards";
+  toggleViewType: () => void;
+  handleAddClick: () => void;
+  handleExportData: () => void;
+}
+
+export const SalesFilters: React.FC<SalesFiltersProps> = ({
+  showFilters,
+  setShowFilters,
+  filters,
+  handleFilterChange,
+  products,
+  salers,
+  clients,
+  fetchSales,
+  viewType,
+  toggleViewType,
+  handleAddClick,
+  handleExportData
+}) => {
+  return (
+    <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-4 md:p-6 mb-6 md:mb-8">
+      <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center gap-4">
+        <form
+          onSubmit={(e) => {
+            e.preventDefault();
+            fetchSales();
+          }}
+          className="relative flex-1 w-full max-w-md"
+        >
+          <input
+            type="text"
+            placeholder="Search sales..."
+            className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+            value={filters.search || ""}
+            onChange={handleFilterChange}
+            name="search"
+          />
+        </form>
+
+        <div className="flex flex-wrap gap-2 w-full sm:w-auto">
+          <button
+            onClick={() => setShowFilters(!showFilters)}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+          >
+            <Filter size={16} className="mr-1 md:mr-2" />
+            <span>Filters</span>
+            {showFilters ? (
+              <ChevronUp size={16} className="ml-1" />
+            ) : (
+              <ChevronDown size={16} className="ml-1" />
+            )}
+          </button>
+          
+          <button
+            onClick={toggleViewType}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+          >
+            <FileText size={16} className="mr-1" />
+            <span>{viewType === "table" ? "Cards" : "Table"}</span>
+          </button>
+          
+          <button
+            onClick={handleExportData}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+          >
+            <Download size={16} className="mr-1" />
+            <span>Export</span>
+          </button>
+          
+          <button
+            onClick={fetchSales}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+          >
+            <RefreshCw size={16} />
+          </button>
+          
+          <button
+            onClick={handleAddClick}
+            className="flex items-center px-3 md:px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm font-medium"
+          >
+            <Plus size={16} className="mr-1 md:mr-2" />
+            <span>New Sale</span>
+          </button>
+        </div>
+      </div>
+
+      {showFilters && (
+        <div
+          id="filters-panel"
+          className="mt-4 p-4 bg-gray-50 rounded-lg border border-gray-200 transition-all"
+        >
+          <h3 className="text-sm font-medium text-gray-700 mb-3 flex items-center">
+            <Filter size={16} className="mr-2" />
+            Filters
+          </h3>
+          <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4">
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Product
+              </label>
+              <select
+                name="productId"
+                value={filters.productId || ""}
+                onChange={handleFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              >
+                <option value="">All Products</option>
+                {products.map((product) => (
+                  <option key={product.id} value={product.id}>
+                    {product.name}
+                  </option>
+                ))}
+              </select>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Saler
+              </label>
+              <select
+                name="salerId"
+                value={filters.salerId || ""}
+                onChange={handleFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              >
+                <option value="">All Salers</option>
+                {salers.map((saler) => (
+                  <option key={saler.id} value={saler.id}>
+                    {saler.name}
+                  </option>
+                ))}
+              </select>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Client
+              </label>
+              <select
+                name="clientId"
+                value={filters.clientId || ""}
+                onChange={handleFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              >
+                <option value="">All Clients</option>
+                {clients.map((client) => (
+                  <option key={client.id} value={client.id}>
+                    {client.name}
+                  </option>
+                ))}
+              </select>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Status
+              </label>
+              <select
+                name="status"
+                value={filters.status || ""}
+                onChange={handleFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              >
+                <option value="">All Statuses</option>
+                <option value="paid">Paid</option>
+                <option value="partial">Partial</option>
+                <option value="unpaid">Unpaid</option>
+              </select>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Start Date
+              </label>
+              <input
+                type="date"
+                name="startDate"
+                value={filters.startDate || ""}
+                onChange={handleFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              />
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                End Date
+              </label>
+              <input
+                type="date"
+                name="endDate"
+                value={filters.endDate || ""}
+                onChange={handleFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              />
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Items per page
+              </label>
+              <select
+                name="pageSize"
+                value={filters.pageSize || 10}
+                onChange={handleFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              >
+                <option value={10}>10</option>
+                <option value={25}>25</option>
+                <option value={50}>50</option>
+                <option value={100}>100</option>
+              </select>
+            </div>
+            <div className="flex items-end">
+              <button
+                onClick={fetchSales}
+                className="px-4 py-2 bg-blue-600 text-white rounded-md text-sm font-medium hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500"
+              >
+                Apply Filters
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
