+++ inventory-management/src/pages/sales/SalesTable.tsx
@@ -1,0 +1,246 @@
+import React from 'react';
+import { Package, Calendar, Eye, Edit2, Trash2 } from 'lucide-react';
+
+interface SalesTableProps {
+  loading: boolean;
+  error: string | null;
+  sales: any[];
+  currentPage: number;
+  pageSize: number;
+  totalPages: number;
+  handlePageChange: (page: number) => void;
+  handleEditClick: (sale: any) => void;
+  handleDeleteConfirm: (id: number) => void;
+  setSelectedSale: (sale: any) => void;
+  setShowViewModal: (show: boolean) => void;
+  sortConfig: any;
+  requestSort: (key: string) => void;
+  getStatusBadge: (sale: any) => React.ReactNode;
+}
+
+export const SalesTable: React.FC<SalesTableProps> = ({
+  loading,
+  error,
+  sales,
+  handleEditClick,
+  handleDeleteConfirm,
+  setSelectedSale,
+  setShowViewModal,
+  sortConfig,
+  requestSort,
+  getStatusBadge
+}) => {
+  const renderSkeleton = () => {
+    return Array(5).fill(0).map((_, i) => (
+      <tr key={`skeleton-${i}`} className="animate-pulse">
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-3/4 mb-2"></div>
+          <div className="h-3 bg-gray-200 rounded w-1/2"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-2/3"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-3/4"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-1/2"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-2/3"></div>
+        </td>
+        <td className="px-6 py-4">
+          <div className="h-4 bg-gray-200 rounded w-1/2"></div>
+        </td>
+        <td className="px-6 py-4 text-right">
+          <div className="h-6 bg-gray-200 rounded w-16 inline-block"></div>
+        </td>
+      </tr>
+    ));
+  };
+
+  return (
+    <div 
+      id="sales-table-container" 
+      className="bg-white rounded-xl shadow-sm border border-gray-200 overflow-hidden"
+    >
+      <div className="overflow-x-auto">
+        <table className="min-w-full divide-y divide-gray-200">
+          <thead className="bg-gray-50">
+            <tr>
+              <th
+                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100"
+                onClick={() => requestSort("saleReference")}
+              >
+                <div className="flex items-center">
+                  Reference
+                  {sortConfig?.key === "saleReference" && (
+                    <span className="ml-1">
+                      {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                    </span>
+                  )}
+                </div>
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Product
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Details
+              </th>
+              <th
+                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100"
+                onClick={() => requestSort("quantity")}
+              >
+                <div className="flex items-center">
+                  Qty (Kg)
+                  {sortConfig?.key === "quantity" && (
+                    <span className="ml-1">
+                      {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                    </span>
+                  )}
+                </div>
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Amount
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Saler
+              </th>
+              <th
+                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100"
+                onClick={() => requestSort("createdAt")}
+              >
+                <div className="flex items-center">
+                  Date
+                  {sortConfig?.key === "createdAt" && (
+                    <span className="ml-1">
+                      {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                    </span>
+                  )}
+                </div>
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Status
+              </th>
+              <th className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Actions
+              </th>
+            </tr>
+          </thead>
+          <tbody className="bg-white divide-y divide-gray-200">
+            {loading ? (
+              renderSkeleton()
+            ) : error ? (
+              <tr>
+                <td
+                  colSpan={9}
+                  className="px-6 py-4 text-center"
+                >
+                  <div className="flex items-center justify-center text-red-600">
+                    {error}
+                  </div>
+                </td>
+              </tr>
+            ) : sales.length === 0 ? (
+              <tr>
+                <td
+                  colSpan={9}
+                  className="px-6 py-4 text-center text-gray-500"
+                >
+                  No sales found.
+                </td>
+              </tr>
+            ) : (
+              sales.map((sale) => (
+                <tr key={sale.id} className="hover:bg-gray-50 transition-colors">
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm font-medium text-gray-900">
+                      {sale.saleReference || "N/A"}
+                    </div>
+                    <div className="text-xs text-gray-500">
+                      {new Date(sale.createdAt).toLocaleDateString()}
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm text-gray-900">
+                      {sale.product?.name || "N/A"}
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm text-gray-900">
+                      <div className="flex items-center">
+                        <Package className="w-4 h-4 mr-1 text-blue-500" />
+                        {sale.quantity} Kg
+                      </div>
+                      <div className="text-xs text-gray-500 mt-1">
+                        @ {sale.unitPrice ? parseFloat(sale.unitPrice).toFixed(2) : "N/A"} RWF/Kg
+                      </div>
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm text-gray-900">
+                      {sale.quantity} Kg
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm text-gray-900">
+                      {sale.unitPrice ? (parseFloat(sale.unitPrice) * sale.quantity) : "N/A"} RWF
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm text-gray-900">
+                      {sale.saler?.user?.profile?.names || "N/A"}
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="flex items-center">
+                      <Calendar className="flex-shrink-0 h-4 w-4 text-gray-400 mr-1" />
+                      <div className="text-sm text-gray-900">
+                        {new Date(sale.createdAt).toLocaleDateString()}
+                      </div>
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    {getStatusBadge(sale)}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                    <div className="flex items-center justify-end space-x-2">
+                      <button
+                        onClick={() => {
+                          setSelectedSale(sale);
+                          setShowViewModal(true);
+                        }}
+                        className="text-green-600 hover:text-green-900 p-1 rounded-full hover:bg-green-50"
+                        title="View Details"
+                      >
+                        <Eye size={18} />
+                      </button>
+                      
+                      <button
+                        onClick={() => handleEditClick(sale)}
+                        className="text-blue-600 hover:text-blue-900 p-1 rounded-full hover:bg-blue-50"
+                        title="Edit Sale"
+                      >
+                        <Edit2 size={18} />
+                      </button>
+                      
+                      <button
+                        onClick={() => handleDeleteConfirm(sale.id)}
+                        className="text-red-600 hover:text-red-900 p-1 rounded-full hover:bg-red-50"
+                        title="Delete Sale"
+                      >
+                        <Trash2 size={18} />
+                      </button>
+                    </div>
+                  </td>
+                </tr>
+              ))
+            )}
+          </tbody>
+        </table>
+      </div>
+
+      {/* Pagination would go here */}
+    </div>
+  );
+};
